ocr/requirements.txt
# Web scraping and browser automation
selenium>=4.15.0
beautifulsoup4>=4.12.0
webdriver-manager>=4.0.0

# PDF processing and OCR
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0

# AI processing
openai>=1.3.0
python-dotenv>=1.0.0

# Progress bars and utilities
tqdm>=4.66.0

# Standard libraries (usually included with Python)
# json, time, datetime, pathlib, typing, hashlib, logging, argparse, re, os, sys, io